afc/sendfile: do not expose idevice_connection_sendfile as public api
