+++ src/idevice.c
@@ -127,20 +127,6 @@ static void __attribute__((destructor)) libimobiledevice_deinitialize(void)
 }
 #endif
 
-/* sendfile support for Apple+Linux.
- * Other OSes fall back to read/write */
-#ifdef __APPLE__
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <sys/uio.h>
-#include <sys/stat.h>
-#elif __linux
-#include <sys/sendfile.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <unistd.h>
-#endif
-
 static idevice_event_cb_t event_cb = NULL;
 
 static void usbmux_event_cb(const usbmuxd_event_t *event, void *user_data)
@@ -356,89 +342,6 @@ idevice_error_t idevice_connection_send(idevice_connection_t connection, const c
 		return IDEVICE_E_SSL_ERROR;
 	}
 	return internal_connection_send(connection, data, len, sent_bytes);
-}
-
-idevice_error_t idevice_connection_sendfile(idevice_connection_t connection, int fd, off_t length, off_t *sent_bytes)
-{
-    struct stat st;
-    int conn_fd;
-
-    if (!connection) {
-        return IDEVICE_E_INVALID_ARG;
-    }
-
-    fstat(fd, &st);
-    if (!S_ISREG(st.st_mode)) {
-        debug_info("ERROR: Unable to use sendfile() input fd is not a regular file");
-        return IDEVICE_E_INVALID_ARG;
-    }
-
-    if (length == 0) {
-        length = st.st_size;
-    }
-
-    conn_fd = (int)(long)connection->data;
-    errno = 0;
-
-    /* TODO: Native sendfile() use on FreeBSD. Currently falls back to read/write.
-     *       It has yet a 3rd interface compared to Apple and Linux.
-     *       See http://www.freebsd.org/cgi/man.cgi?query=sendfile&sektion=2 */
-
-#ifdef __APPLE__
-    /* int sendfile(int fd, int s, off_t offset, off_t *len, struct sf_hdtr *hdtr, int flags); */
-    off_t len = length;
-    int result = sendfile(fd, conn_fd, 0, &len, NULL, 0);
-    *sent_bytes = len;
-    if (result == 0) {
-        return IDEVICE_E_SUCCESS;
-    } else {
-        debug_info("ERROR: When calling sendfile(%d, %d, 0, %d) __APPLE__: %s", fd, conn_fd, length, strerror(errno));
-        return IDEVICE_E_UNKNOWN_ERROR;
-    }
-#elif __linux
-    /* ssize_t sendfile(int out_fd, int in_fd, off_t * offset, size_t count); */
-    ssize_t len = sendfile(conn_fd, fd, NULL, length);
-    *sent_bytes = len;
-    if (len >= 0) {
-        return IDEVICE_E_SUCCESS;
-    } else {
-        debug_info("ERROR: When calling sendfile(%d, %d, NULL, %d) __linux: %s", conn_fd, fd, length, strerror(errno));
-        return IDEVICE_E_UNKNOWN_ERROR;
-    }
-#else
-    /* 10 pages of memory gives close to optimum performance,
-     * while not being too aggresive on memory consumption */
-    char     stack_buf[4096 * 10];
-    char    *buf = &stack_buf[0];
-    ssize_t  num_read;
-
-    while ((num_read = read(fd, (void *) buf, sizeof(stack_buf))) > 0) {
-
-        ssize_t num_written = 0;
-        while ((num_written = write(conn_fd, (void *) (buf + num_written), num_read)) > 0) {
-            *sent_bytes += num_written;
-            num_read -= num_written;
-            if (num_read == 0) {
-                break;
-            } else if (num_read < 0) {
-                debug_info("ERROR: internal error. Wrote more than we read!");
-                return IDEVICE_E_UNKNOWN_ERROR;
-            }
-        }
-        if (num_written < 0) {
-           debug_info("ERROR: write(%d,,) failed: %s", conn_fd, strerror(errno));
-           return IDEVICE_E_UNKNOWN_ERROR;
-        }
-    }
-
-    if (num_read == 0 && *sent_bytes == length) {
-        return IDEVICE_E_SUCCESS;
-    } else {
-        /* must be a read error, write errors return from inner while loop */
-        debug_info("ERROR: read(%d,,): %s", fd, strerror(errno));
-        return IDEVICE_E_UNKNOWN_ERROR;
-    }
-#endif
 }
 
 /**
